(fix) free of mlx pointer
Final free of mlx pointer was missing at different places. mlx_destroy_window() only destroys the connection to X11 but the pointer to mlx itself still needs to be freed. Added free of mlx pointer to init_mlx_interface() in window_mgmt.c and to cleanup_mlx() in cleanup.c